backend/requirements.txt
# Web Framework
flask==3.0.0
flask-cors==4.0.0

# LangChain and AI
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5
openai>=1.10.0,<2.0.0

# Excel/Data Processing
pandas==2.1.4
openpyxl==3.1.2
xlrd==2.0.1

# HTTP and API
requests==2.31.0

# Environment and Configuration
python-dotenv==1.0.0

# Redis for caching
redis==5.0.1

# Clerk authentication
pyjwt==2.8.0
cryptography==42.0.4

# Development and Testing
pytest==7.4.4
pytest-flask==1.3.0